TTPD-174: moves caching to the separate package mapping/cache
