+++ mapping/mapping.go
@@ -5,26 +5,21 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
-	"time"
+
+	"github.com/Shop2market/go-client/mapping/cache"
 )
 
-const (
-	PATH      = "/api/v1/mapping_files.json"
-	CACHE_TTL = 1 * time.Hour
-)
+const PATH = "/api/v1/mapping_files.json"
 
 type creds struct {
 	Endpoint string
 	Username string
 	Password string
 }
-type Cache struct {
-	Data map[string][][]string
-	Date *time.Time
-}
+
 type Repo struct {
 	creds
-	Cache
+	cache.Cache
 }
 
 func New(endpoint, username, password string) (repo *Repo, err error) {
@@ -33,7 +28,7 @@ func New(endpoint, username, password string) (repo *Repo, err error) {
 		return
 	}
 	creds := creds{Endpoint: endpoint, Username: username, Password: password}
-	Cache := Cache{make(map[string][][]string), nil}
+	Cache := cache.Cache{make(map[string][][]string), nil}
 	repo = &Repo{creds, Cache}
 	return
 }
@@ -85,23 +80,4 @@ func (repo *Repo) prepareRequest() (request *http.Request, err error) {
 	}
 	request.SetBasicAuth(repo.Username, repo.Password)
 	return
-}
-
-func (c *Cache) Update(mappings map[string][][]string) {
-	now := time.Now().UTC()
-	c.Data = mappings
-	c.Date = &now
-}
-
-func (c *Cache) IsValid() bool {
-	if c.IsEmpty() {
-		return false
-	}
-	now := time.Now().UTC()
-	expiration := c.Date.Add(CACHE_TTL)
-	return expiration.After(now)
-}
-
-func (c *Cache) IsEmpty() bool {
-	return c.Data == nil || c.Date == nil
 }
